Refactor mailing list updates into separate interactor
This will allow us to just call interactors in bulk for an initial
import of data, and divides responsibility a bit more cleanly.
