+++ app/interactors/update_profile.rb
@@ -45,25 +45,12 @@ class UpdateProfile
   end
 
   def update_person_on_mailing_list
-    return unless Rails.configuration.mailing_list_integration_enabled
-
-    UpdatePersonOnMailingListWorker.perform_async(
-      person.email,
-      @email_before_update,
-      person.given_name,
-      person.surname,
-      mailchimp_tags
-    )
+    UpdatePersonOnMailingList.call(person: person, previous_email: @email_before_update)
   end
 
   def touch_person_last_edited_or_confirmed_at
     # We don't mind not running validations on touching this attribute - this call should not fail
     person.touch(:last_edited_or_confirmed_at) # rubocop:disable Rails/SkipsModelValidations
-  end
-
-  def mailchimp_tags
-    # TODO: This is an MVP and needs to move into a more appropriate place
-    ['pf_imported'] + person.building.select(&:present?).map { |building| "pf_building_#{building}" }
   end
 
   def person
