+++ spec/interactors/update_profile_spec.rb
@@ -9,7 +9,6 @@ describe UpdateProfile do
       name: 'Per Son',
       given_name: 'Per',
       surname: 'Son',
-      building: ['', 'skyscraper', 'airport'],
       assign_attributes: true,
       save!: true,
       valid?: valid,
@@ -20,13 +19,11 @@ describe UpdateProfile do
   let(:person_attributes) { double('Attributes') }
   let(:notifier) { instance_double(GovukNotify, updated_profile: true) }
   let(:notify_of_change) { false }
-  let(:mailing_list_integration_enabled) { true }
 
   before do
     allow(person).to receive(:notify_of_change?).with(instigator).and_return(notify_of_change)
     allow(person).to receive(:email).and_return('person@gov.uk', 'person@example.com')
-    allow(UpdatePersonOnMailingListWorker).to receive(:perform_async)
-    allow(Rails.configuration).to receive(:mailing_list_integration_enabled).and_return(mailing_list_integration_enabled)
+    allow(UpdatePersonOnMailingList).to receive(:call)
   end
 
   describe '.call' do
@@ -58,6 +55,10 @@ describe UpdateProfile do
       it 'does not send an email to the updated user' do
         expect(notifier).not_to have_received(:updated_profile)
       end
+
+      it 'does not update the user on the mailing list' do
+        expect(UpdatePersonOnMailingList).not_to have_received(:call)
+      end
     end
 
     context 'when the person is valid' do
@@ -76,26 +77,12 @@ describe UpdateProfile do
         expect(notifier).not_to have_received(:updated_profile)
       end
 
-      it 'queues an UpdatePersonOnMailingListWorker with the previous email' do
-        expect(UpdatePersonOnMailingListWorker).to have_received(:perform_async).with(
-          'person@example.com',
-          'person@gov.uk',
-          'Per',
-          'Son',
-          %w[pf_imported pf_building_skyscraper pf_building_airport]
-        )
-      end
-
       it 'touches the last_edited_or_confirmed_at timestamp' do
         expect(person).to have_received(:touch).with(:last_edited_or_confirmed_at)
       end
 
-      context 'when mailing list integration is disabled' do
-        let(:mailing_list_integration_enabled) { false }
-
-        it 'does not queue an UpdatePersonOnMailingListWorker' do
-          expect(UpdatePersonOnMailingListWorker).not_to have_received(:perform_async)
-        end
+      it 'updates the person on the mailing list' do
+        expect(UpdatePersonOnMailingList).to have_received(:call).with(person: person, previous_email: 'person@gov.uk')
       end
 
       context 'when the updated person should be notified of changes' do
